+++ tests/domain/test_batch_domain.py
@@ -19,6 +19,7 @@ from shotgrid_leecher.record.shotgrid_subtypes import (
     AssetFieldsMapping,
     ShotFieldsMapping,
     TaskFieldsMapping,
+    StepFieldsMapping,
 )
 from shotgrid_leecher.writers import db_writer
 
@@ -38,6 +39,7 @@ def _default_fields_mapping() -> FieldsMapping:
         AssetFieldsMapping.from_dict({}),
         ShotFieldsMapping.from_dict({}),
         TaskFieldsMapping.from_dict({}),
+        StepFieldsMapping.from_dict({}),
     )