+++ tests/utils/test_collections.py
@@ -45,6 +45,31 @@ def test_flatten_dict_nested_non_dicts():
     assert_that(actual.get(f"{k1}.{k2}.{k3}")).is_equal_to(k3.int)
 
 
+def test_flatten_dict_with_except():
+    # Arrange
+    def _f():
+        return str(uuid.uuid4())[:5]
+
+    k1, k2, k3 = _f(), _f(), _f()
+    sub_dict = {1: 1, 2: 2, 3: 3}
+    data = {k1: {k2: {k3: sub_dict}}, k3: {k2: {k1: sub_dict}}}
+    # Act
+    actual = flatten_dict(data, {f"{k1}.{k2}.{k3}"})
+    # Assert
+    assert_that(set(actual.keys())).is_equal_to(
+        {
+            f"{k1}.{k2}.{k3}",
+            f"{k3}.{k2}.{k1}.1",
+            f"{k3}.{k2}.{k1}.2",
+            f"{k3}.{k2}.{k1}.3",
+        }
+    )
+    assert_that(actual[f"{k1}.{k2}.{k3}"]).is_equal_to(sub_dict)
+    assert_that(actual[f"{k3}.{k2}.{k1}.1"]).is_equal_to(1)
+    assert_that(actual[f"{k3}.{k2}.{k1}.2"]).is_equal_to(2)
+    assert_that(actual[f"{k3}.{k2}.{k1}.3"]).is_equal_to(3)
+
+
 def test_flatten_dict_nested_mixed_dicts():
     # Arrange
     k1, k2, k3 = uuid.uuid4(), uuid.uuid4(), uuid.uuid4()